+++ src/app/clusters/general-commissioning-server/app_config_dependent_sources.cmake
@@ -13,8 +13,19 @@
 # limitations under the License.
 
 # This is the equivalent to app_config_dependent_sources.gni
+
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
-    "${CLUSTER_DIR}/general-commissioning-server.cpp"
+    "${CLUSTER_DIR}/CodegenIntegration.cpp"
+)
+
+# These are the things that BUILD.gn dependencies would pull
+TARGET_SOURCES(
+  ${APP_TARGET}
+  PRIVATE
+    "${CLUSTER_DIR}/general-commissioning-cluster.cpp"
+    "${CLUSTER_DIR}/general-commissioning-cluster.h"
+    "${CLUSTER_DIR}/general-commissioning-logic.cpp"
+    "${CLUSTER_DIR}/general-commissionings-logic.h"
 )
