+++ src/app/clusters/general-commissioning-server/general-commissioning-cluster.h
@@ -1,0 +1,58 @@
+/*
+ *
+ *    Copyright (c) 2024-25 Project CHIP Authors
+ *    All rights reserved.
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+#pragma once
+
+#include <app/clusters/general-commissioning-server/general-commissioning-logic.h>
+#include <app/server-cluster/DefaultServerCluster.h>
+#include <clusters/GeneralCommissioning/ClusterId.h>
+#include <protocols/interaction_model/StatusCode.h>
+
+namespace chip {
+namespace app {
+namespace Clusters {
+
+class GeneralCommissioningCluster : public DefaultServerCluster
+{
+public:
+    GeneralCommissioningCluster(EndpointId endpointId, BitFlags<GeneralCommissioning::Feature> featureFlags) :
+        DefaultServerCluster({ endpointId, GeneralCommissioning::Id }), mLogic(featureFlags)
+    {}
+
+    DataModel::ActionReturnStatus ReadAttribute(const DataModel::ReadAttributeRequest & request,
+                                                AttributeValueEncoder & encoder) override;
+
+    std::optional<DataModel::ActionReturnStatus> InvokeCommand(const DataModel::InvokeRequest & request,
+                                                               TLV::TLVReader & input_arguments, CommandHandler * handler) override;
+
+    CHIP_ERROR AcceptedCommands(const ConcreteClusterPath & path,
+                                ReadOnlyBufferBuilder<DataModel::AcceptedCommandEntry> & builder) override;
+
+    CHIP_ERROR GeneratedCommands(const ConcreteClusterPath & path, ReadOnlyBufferBuilder<CommandId> & builder) override;
+
+    CHIP_ERROR Attributes(const ConcreteClusterPath & path, ReadOnlyBufferBuilder<DataModel::AttributeEntry> & builder) override;
+
+    GeneralCommissioningLogic & GetLogic() { return mLogic; }
+
+private:
+    GeneralCommissioningLogic mLogic;
+};
+
+} // namespace Clusters
+} // namespace app
+} // namespace chip
